+++ components.d.ts
@@ -30,6 +30,10 @@ declare module '@vue/runtime-core' {
     BaseFormTwo: typeof import('./src/components/BaseFormTwo.vue')['default']
     BaseInput: typeof import('./src/components/BaseInput.vue')['default']
     BaseMessage: typeof import('./src/components/BaseMessage.vue')['default']
+    BaseQuestionAll: typeof import('./src/components/BaseQuestionAll.vue')['default']
+    BaseQuestionScore: typeof import('./src/components/BaseQuestionScore.vue')['default']
+    BaseQuestionText: typeof import('./src/components/BaseQuestionText.vue')['default']
+    BaseQuestionwll: typeof import('./src/components/BaseQuestionwll.vue')['default']
     BaseScore: typeof import('./src/components/BaseScore.vue')['default']
     BaseSelect: typeof import('./src/components/BaseSelect.vue')['default']
     BaseSvgIcon: typeof import('./src/components/BaseSvgIcon.vue')['default']
@@ -37,6 +41,7 @@ declare module '@vue/runtime-core' {
     BaseTag: typeof import('./src/components/BaseTag.vue')['default']
     BaseTextarea: typeof import('./src/components/BaseTextarea.vue')['default']
     BaseToast: typeof import('./src/components/BaseToast.vue')['default']
+    Baswll: typeof import('./src/components/Baswll.vue')['default']
     EmployeeLeaderCreate: typeof import('./src/components/EmployeeLeaderCreate.vue')['default']
     EmployeeLeaderMain: typeof import('./src/components/EmployeeLeaderMain.vue')['default']
     EmployeeLeaderUpdate: typeof import('./src/components/EmployeeLeaderUpdate.vue')['default']
