+++ src/components/BaseQuestionAll.vue
