+++ src/components/BaseQuestionScore.vue
